feat #5 add user graphQL resolvers, typeDefs
