+++ graphql/_queries.js
@@ -1,0 +1,9 @@
+const { gql } = require('apollo-server');
+
+const typeDefs = gql`
+  type Query {
+    users: [User]
+  }
+`;
+
+module.exports = typeDefs;
