--- graphql/resolvers/index.js
@@ -1,9 +1,0 @@
-const userResolvers = require('./user');
-
-const resolvers = {
-  Query: {
-    ...userResolvers.Query,
-  },
-};
-
-module.exports = resolvers;
